blue-chip-screener/requirements.txt
yfinance
pandas
plotly
pyyaml
requests
lxml
jinja2
